+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0.22] - 2026-02-25
+
+### Added
+
+- Gzip and Brotli compression support for merged appsettings files. After merging, `.gz` and `.br` versions are automatically created for both `appsettings.json` and `appsettings.{environment}.json` files using System.IO.Compression (#401)
+
+### Fixed
+
+- Only delete Brotli and Gzip files with SettingsFileType.Environment, preserving Primary type compressed files (#399)
+
 ## [0.3.4.21] - 2026-02-11
 
 ## [0.3.3.20] - 2026-01-16
@@ -81,7 +91,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.3.4] - 2023-11-17
 
-[unreleased]: https://github.com/Afterlife-Guide/AppSettings.Merge/compare/0.3.4.21...HEAD
+[unreleased]: https://github.com/Afterlife-Guide/AppSettings.Merge/compare/0.4.0.22...HEAD
+[0.4.0.22]: https://github.com/Afterlife-Guide/AppSettings.Merge/compare/0.3.4.21...0.4.0.22
 [0.3.4.21]: https://github.com/Afterlife-Guide/AppSettings.Merge/compare/0.3.3.20...0.3.4.21
 [0.3.3.20]: https://github.com/Afterlife-Guide/AppSettings.Merge/compare/0.3.2.19...0.3.3.20
 [0.3.2.19]: https://github.com/Afterlife-Guide/AppSettings.Merge/compare/0.3.1.18...0.3.2.19